Add support for using `dup2` to replace stdin/stdout file descriptors.
One of the common use cases for `dup2` is to use it to replace stdin and
stdout with new file descriptors. Add support for this, and add a test
that does this.

This adds new `take_stdin`, `take_stdout`, and `take_stderr` functions
which return `OwnedFd`, allowing it to be passed into `dup2`.