Check in pre-assembled libraries for the outline asm code.
Check in pre-assembled libraries for the outline asm code, allowing the
"cc" dependency to be optional. This allows rsix to be built on hosts
that don't have a C compiler installed.

`feature = "cc"` enables the use of the `cc` package, and builds the
libraries from source, including checking that the built versions match
the checked-in versions, and is enabled in CI.

Fixes #78.