Rename `Signal::Alrm` to `Signal::Alarm`, and several others.
Following a suggestion by @cgwalters in #177, which is in keeping with
rustix's overall [philosophy] of using more human-friendly names, rename several
`Signal` arms to re-insert missing inner vowels.

This leaves some abbreviations in place, such as `Segv`; I'm open to
suggestions here, but offhand it feels like these are sufficiently
recognizable and convenient that they can stay as they are.

[philosophy]: https://docs.rs/rustix/latest/rustix/#