Updates for rustc-dep-of-std mode.
In rustc-dep-of-std mode, within a std build, `can_compile` runs at
a time when it doesn't have access to `std` or even `core` yet. So
avoid using it to test for `asm!` support.

And, avoid depending on `libc`, even if just to enable its
rustc-dep-of-std build, because something about the way this works
doesn't work with std's build system.