Detect when `statx` isn't available on older versions of Docker.
In older versions of Docker, `statx` can fail with `EPERM` meaning it
isn't available. Recognize this, and translate it into `ENOSYS`. Also,
have `statx` remember this and fail quickly in future calls.

This is more emulation than rustix usually likes to do, but in this
case, this is the behavior seems likely to be what most users will
want, and we might as well put the code in one place.

Fixes #352.